Refactored everything for better organization, removed clautod references from log.py
