+++ devscripts/release.sh
@@ -49,10 +49,11 @@ fi
 
 # Tag the current commit in git
 echo "[release] Package built. Tagging current commit as v"$1
-if ! git tag -a v$1 && git push --tags ; then
+if ! git tag -a v$1 ; then
     echo "[release] Failed to tag commit. This is not a legitimate release!"
     exit 1
 fi
+git push --tags
 
 echo "[release] Commit tagged. Publishing to local Debian repo..."
